moving /link test cases to test_link_router
